+++ src/features/admin/adminRouter.ts
@@ -1,10 +1,11 @@
 import express from "express";
-import { getAllUsers, confirmUser, setVendorStatus } from "./adminController";
+import { getAllUsers, confirmUser, setVendorStatus, getAllOrders } from "./adminController";
 
 const router = express.Router();
 
 router.put("/users/:userId/confirm", confirmUser);
 router.put("/users/:userId/vendor-status", setVendorStatus);
 router.get("/users", getAllUsers);
+router.get("/all-orders", getAllOrders);
 
 export default router;
